refactor: consolidate atlas class into a single shared class, re-usable for other atlases
