+++ tests/test_models.py
@@ -8,7 +8,7 @@ from aind_data_schema_models.platforms import Platform
 from aind_data_schema_models.registries import Registry
 from aind_data_schema_models.species import Species
 from aind_data_schema_models.mouse_anatomy import MouseAnatomicalStructure
-from aind_data_schema_models.mouse_ccf import CCFStructure
+from aind_data_schema_models.brain_atlas import CCFStructure
 
 
 class LiteralAndDefaultTests(unittest.TestCase):
